+++ Automation/CHANGELOG.md
@@ -2,6 +2,7 @@
 * [Feature] Add debug signals for district stock tracking: `Debug.DistrictStockTracker.<GoodId>`.
 * [Feature] Add `Collectable.Ready` signal to check if the building can collect any gatherable or cuttable resource.
 * [Feature] Add templates to check if a gathering building can collect any gatherable or cuttable resource.
+* [Feature] Add `Plantable.Ready` signal that tells how many plantable resources can be planted in the range.
 * [Fix] In "describe mode," show action argument values as real numbers instead of script fixed point values.
 
 # v2.6.0 (June 22nd, 2025):